# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Topic Service API
  description: 아지트 토픽 API
  version: v1.0.0
servers:
- url: http://localhost
  description: Generated server url
tags:
- name: Topic
  description: 토픽(주제) API. 영상 격자는 /videos 하위 리소스.
paths:
  /api/v1/topics:
    get:
      tags:
      - Topic
      summary: 토픽 목록 조회
      description: "아지트의 startAt 최신 토픽 최대 10개. 영상 격자는 GET /{topicUuid}/videos. actor가\
        \ 있으면 uploadedByMe를 채웁니다."
      operationId: list
      parameters:
      - name: agitUuid
        in: query
        description: 아지트 UUID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicResponseDto"
    post:
      tags:
      - Topic
      summary: 토픽 생성
      description: 아지트에 주제를 만듭니다. 생성자는 X-User-UUID이며 ACTIVE 멤버만 가능합니다. 영상은 붙이지 않습니다.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTopicRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TopicResponseDto"
  /api/v1/topics/{topicUuid}/videos:
    get:
      tags:
      - Topic
      summary: 토픽 영상 목록
      description: 주제 상세 격자.
      operationId: listVideos
      parameters:
      - name: topicUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicVideoResponseDto"
    post:
      tags:
      - Topic
      summary: 토픽에 영상 붙이기
      description: 사용자당 토픽당 1개. 같은 영상 재요청은 200. 업로더는 actor이며 ACTIVE 멤버만 가능합니다.
      operationId: attachVideo
      parameters:
      - name: topicUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachTopicVideoRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TopicVideoResponseDto"
  /api/v1/topics/{topicUuid}:
    get:
      tags:
      - Topic
      summary: 토픽 단건 조회
      operationId: get
      parameters:
      - name: topicUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TopicResponseDto"
    delete:
      tags:
      - Topic
      summary: 토픽 삭제
      description: 영상이 없는 토픽만 소프트 삭제합니다. 생성자 또는 HOST만 가능합니다.
      operationId: delete
      parameters:
      - name: topicUuid
        in: path
        description: 토픽 UUID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: No Content
    patch:
      tags:
      - Topic
      summary: 토픽 수정
      description: 제목 또는 진행일. 전달하지 않은 필드는 유지됩니다. 생성자 또는 HOST만 가능합니다.
      operationId: update
      parameters:
      - name: topicUuid
        in: path
        description: 토픽 UUID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTopicRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TopicResponseDto"
  /api/v1/topics/list:
    get:
      tags:
      - Topic
      summary: 토픽 구간 목록 조회
      description: 아지트의 토픽을 KST 날짜 기준 ONGOING/UPCOMING/PAST로 조회합니다. 최신 10개 갤러리는 GET
        /topics. actor가 있으면 uploadedByMe를 채웁니다.
      operationId: listByStatus
      parameters:
      - name: agitUuid
        in: query
        description: 아지트 UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        description: "KST 날짜 구간. ONGOING=오늘, UPCOMING=이후, PAST=이전"
        required: true
        schema:
          type: string
          enum:
          - ONGOING
          - UPCOMING
          - PAST
      - name: limit
        in: query
        description: "최대 개수. 생략 시 10, 1~20으로 제한"
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicResponseDto"
  /api/v1/topics/calendar:
    get:
      tags:
      - Topic
      summary: 토픽 캘린더 조회
      description: 해당 연월에서 영상이 있는 날짜만 반환합니다.
      operationId: calendar
      parameters:
      - name: agitUuid
        in: query
        description: 아지트 UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: yearMonth
        in: query
        description: 연월 (yyyy-MM)
        required: true
        schema:
          type: string
        example: 2026-08
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TopicCalendarResponse"
  /api/v1/topics/{topicUuid}/videos/{videoUuid}:
    delete:
      tags:
      - Topic
      summary: 토픽에서 영상 제거
      description: 본인 영상만 제거할 수 있습니다. 업로더는 actor입니다.
      operationId: detachVideo
      parameters:
      - name: topicUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: videoUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "503":
          description: Service Unavailable
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: No Content
components:
  schemas:
    CreateTopicRequest:
      type: object
      description: "토픽 생성 요청. 주제만 만든다. 영상은 POST /topics/{topicUuid}/videos"
      properties:
        agitUuid:
          type: string
          format: uuid
          description: 아지트 UUID
        title:
          type: string
          description: 토픽 제목
          example: 점심 메뉴
        startAt:
          type: string
          format: date-time
          description: 진행 날짜. 없으면 오늘 00:00
      required:
      - agitUuid
    TopicResponseDto:
      type: object
      description: "토픽(주제) 응답. 영상 격자는 GET /topics/{topicUuid}/videos"
      properties:
        topicUuid:
          type: string
          format: uuid
          description: 토픽 UUID
        agitUuid:
          type: string
          format: uuid
          description: 아지트 UUID
        creatorUuid:
          type: string
          format: uuid
          description: 생성자 UUID
        title:
          type: string
          description: 토픽 제목
        startAt:
          type: string
          format: date-time
          description: 진행 날짜
        videoCount:
          type: integer
          format: int32
          description: 붙은 영상 수
        uploadedByMe:
          type: boolean
          description: 조회한 사용자가 이 토픽에 영상을 올렸는지. actor 헤더가 없으면 null
        createdAt:
          type: string
          format: date-time
          description: 생성 시각
    AttachTopicVideoRequest:
      type: object
      description: 기존 토픽에 영상 붙이기. 사용자당 토픽당 1개
      properties:
        videoUuid:
          type: string
          format: uuid
          description: 비디오 UUID
      required:
      - videoUuid
    TopicVideoResponseDto:
      type: object
      description: 토픽에 붙은 영상
      properties:
        videoUuid:
          type: string
          format: uuid
          description: 비디오 UUID
        userUuid:
          type: string
          format: uuid
          description: 업로더 UUID
        createdAt:
          type: string
          format: date-time
          description: 붙은 시각
    UpdateTopicRequest:
      type: object
      description: 토픽 수정 요청. 전달하지 않은 필드는 유지됩니다.
      properties:
        title:
          type: string
          description: 토픽 제목
          example: 저녁 메뉴
        startAt:
          type: string
          format: date-time
          description: 진행 날짜
    TopicCalendarResponse:
      type: object
      description: 토픽 캘린더. 영상이 있는 날짜만 활성이다. 아지트 제목/멤버는 아지트 서비스 스냅샷이며 이 응답에 포함하지 않는다.
      properties:
        agitUuid:
          type: string
          format: uuid
          description: 아지트 UUID
        yearMonth:
          type: string
          description: 조회 연월
          example: 2026-08
        activeDates:
          type: array
          description: 영상이 있는 날짜
          items:
            type: string
            format: date